Renamed attributes and moved class
