+++ applications/harvester-api/src/main/java/no/dcat/harvester/crawler/CatalogHarvestRecord.java
@@ -1,4 +1,4 @@
-package no.dcat.harvester;
+package no.dcat.harvester.crawler;
 
 import lombok.Data;
