+++ applications/harvester-api/src/main/java/no/dcat/harvester/crawler/DatasetHarvestRecord.java
@@ -8,7 +8,7 @@ import java.util.Date;
 @Data
 public class DatasetHarvestRecord {
     private Date date;
-    private String catalogHarvestRecordId;
+    private String datasourceId;
     private String datasetUri;
     private String datasetId;
     private ValidationStatus validationStatus;
